requirements.txt
discord.py>=2.3.2
langchain_ollama